refactor headerfile
