+++ .gitignore
@@ -1,4 +1,5 @@
 .vscode/
+build/
 tests/
 *.exe
 *.txt
