+++ include/error.h
@@ -1,0 +1,11 @@
+#ifndef ERROR_H
+#define ERROR_H
+
+#include <stdio.h>
+
+/*
+ *
+ */
+void error_info(const char *msg, size_t errnum, size_t status);
+
+#endif // ERROR_H
