+++ include/stb_image/stb_image_resize.h
