+++ include/utils.h
@@ -1,0 +1,14 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+/*
+ *
+ */
+void reverse_string(char *str);
+
+/*
+ *
+ */
+void show_usage(void);
+
+#endif // UTILS_H
